+++ app/src/main/java/com/example/appganaderosv1/insert_new_sales.java
@@ -394,9 +394,9 @@ public class insert_new_sales extends AppCompatActivity {
             ventaDetalle.setId_venta_detalle(cursor.getInt(5));
             ventaDetalle.setId_ganado(cursor.getInt(6));
             ventaDetalle.setPeso_canal_venta(cursor.getDouble(7));
-            ventaDetalle.setPrecio_venta(cursor.getInt(8));
+            ventaDetalle.setPrecio_venta(cursor.getDouble(8));
             ventaDetalle.setTara_venta(cursor.getInt(9));
-            ventaDetalle.setTotal_venta(cursor.getInt(10));
+            ventaDetalle.setTotal_venta(cursor.getDouble(10));
 
             compraDetalle = new CompraDetalle();
             compraDetalle.setId_compra_detalle(cursor.getInt(11));
@@ -541,9 +541,9 @@ public class insert_new_sales extends AppCompatActivity {
             ventaDetalle.setId_venta_detalle(cursor.getInt(5));
             ventaDetalle.setId_ganado(cursor.getInt(6));
             ventaDetalle.setPeso_canal_venta(cursor.getDouble(7));
-            ventaDetalle.setPrecio_venta(cursor.getInt(8));
+            ventaDetalle.setPrecio_venta(cursor.getDouble(8));
             ventaDetalle.setTara_venta(cursor.getInt(9));
-            ventaDetalle.setTotal_venta(cursor.getInt(10));
+            ventaDetalle.setTotal_venta(cursor.getDouble(10));
             ventaDetalle.setId_venta(cursor.getInt(11));
 
             compraDetalle = new CompraDetalle();
